+++ src/screens/dashboard/DashboardMain.tsx
@@ -105,6 +105,8 @@ export const DashboardMain = () => {
     return undefined;
   }, [chains]);
 
+  console.log({ publicKey, chains, assets });
+
   return (
     <div className="flex flex-col break-words">
       <div className="grid grid-cols-[auto,1fr,auto] gap-2 mb-6 items-center">
